requirements.txt
# Orchestration
langgraph
langchain
langchain-core

# LLMs
langchain-google-genai

# Vector Database (RAG Memory)
chromadb
langchain-chroma

# Text Splitting (for PDF chunking into ChromaDB)
langchain-text-splitters
langchain-community

# PDF Processing
PyMuPDF

# Environment & Utilities
python-dotenv
certifi

# Web Framework
fastapi
uvicorn[standard]
python-multipart
pydantic
